Add tests to types package
