+++ packages/types/src/index.ts
@@ -4,3 +4,4 @@ export * from './Account';
 export * from './CoinType';
 export * from './Cosmos';
 export * from './FiatCoinType';
+export * from './crypto';
